pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-insight"
version = "0.1.0"
description = "Telemetry-to-Insight pipeline for robotics and autonomous systems"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=12.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "matplotlib>=3.7",
    "torch>=2.0",
    "transformers>=4.35",
    "pytest>=7.0",
]

[project.optional-dependencies]
gpu = ["cudf-cu12>=24.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "data*"]
